requirements.txt
openai
python-dotenv
pandas
psycopg2-binary
numpy
faiss-cpu
requests
beautifulsoup4
